qwen3, inflect2: polish after review
Qwen3: a require helper like the siblings, mandatory keys read with getValue
so a missing one is named, the overflow guard folded into the dimension
check, and requireComplete (lengths sum to the packed token count) once per
ingest instead of a per-chunk scan. Inflect2: the espeak timeout also bounds
the stdout drain, ipaRun is one method, the lexicon's smallest-entry divisor
is explained, the stub tests use @TempDir through one helper, and the
provider test lives in its own class.

+++ jinfer/jinfer-inflect2/src/main/java/com/qxotic/jinfer/models/inflect2/frontend/LexiconPhonemizer.java
@@ -243,11 +243,12 @@ final class LexiconPhonemizer implements Phonemizer {
             long entriesAt = buffer.getLong();
             if (entriesAt < 0 || entriesAt > data.length)
                 throw new IOException("invalid lexicon entry offset " + entriesAt);
+            // shared + fresh + symbols: the smallest entry is three bytes
             if (count > (data.length - entriesAt) / 3)
                 throw new IOException("lexicon entry count exceeds the file size");
 
             Map<String, int[]> lexicon = HashMap.newHashMap(count);
-            buffer.position(Math.toIntExact(entriesAt));
+            buffer.position((int) entriesAt);
             byte[] word = new byte[MAX_WORD];
             int length = 0;
             for (int entry = 0; entry < count; entry++) {
